% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\docType{data}
\name{frequencies}
\alias{frequencies}
\title{Example Data: frequencies}
\format{
## `frequencies`
A list with 1 objects:
\describe{
  \item{frequencies}{A dataframe with 24487 rows and 1 column, each row
  is a maf for a SNP}
}
}
\usage{
frequencies
}
\description{
Minor allele frequencies calculated from the real data.
}
\keyword{datasets}
