% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceComponentFunc.R
\name{geneticEffect}
\alias{geneticEffect}
\alias{noiseEffect}
\title{Simulate the matrix of genetic effects or noise effects}
\usage{
geneticEffect(M, K, W = NULL, V.g = NULL, method = "eigen")

noiseEffect(N, K, V.e = NULL, method = "eigen")
}
\arguments{
\item{M}{The number of variants in a region to simulate.}

\item{K}{The number of phenotypes to simulate.}

\item{W}{The weight assigned to each variant, default is 1.}

\item{V.g}{The covariance matrix among columns, default is 0.5 for
non-diagnal element.}

\item{method}{String specifying the matrix decomposition used to determine
the matrix root of sigma. Possible methods are eigenvalue decomposition
("eigen", default), singular value decomposition ("svd"), and
Cholesky decomposition ("chol"). The Cholesky is typically fastest,
not by much though.}

\item{N}{The number of samples to simulate.}

\item{V.e}{The covariance matrix among rows, default is 0.5 for
each non-diagnal element.}
}
\value{
\code{geneticEffect():} Matrix of genetic effects, rows correspond
   to variants, columns correspond to phenotypes. \code{noiseEffect():}
   Matrix of noise term, rows correspond to sample, columns correspond to
   phenotypes.
}
\description{
Simulate the matrix of genetic effects or noise effects
}
\examples{
out <- geneticEffect(20, 10)
out <- noiseEffect(20, 5)
}
