% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWASfunc.R
\name{GWASfunc}
\alias{GWASfunc}
\title{GWAS summary statistics calculation}
\usage{
GWASfunc(X, Y)
}
\arguments{
\item{X}{Genotypes dataframe list with K objects}

\item{Y}{Phenotype dataframe list with K objects}
}
\value{
GWAS summary statistics dataframe
}
\description{
GWAS summary statistics calculation
}
\examples{
K <- 5; N<- c(100,200,300,150,200); Ns<-100; NrSNP=1000
h2s <- rep(0.6,K); pcorr <- 0.5
out <- PheSimulator(K, N, NrSNP, Ns, pcorr=pcorr, h2s=h2s)
X <- out$Genotype
Y <- out$Phenotype
out <- GWASfunc(X,Y)
}
