% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotypeFunc.R
\name{PheSimulator}
\alias{PheSimulator}
\title{Phenotype simulation.}
\usage{
PheSimulator(
  K,
  N,
  NrSNP,
  Ns,
  genoMethod = "SNPfrequency",
  SNPfrequency = NULL,
  cSNP = 0.01,
  genVar = NULL,
  noiseVar = Null,
  pcorr = NULL,
  pgenVar = 0.6,
  h2s = NULL,
  theta = 0.8,
  pTraits = 1,
  Haplotype,
  ...
)
}
\arguments{
\item{K}{Number [integer] of phenotypes to simulate.}

\item{N}{Vector of sample sizes for K phenotypes. If input is a single integer
which means all phenotypes have same sample size.}

\item{NrSNP}{Number [integer] of genetic variants to simulate.}

\item{Ns}{The number of overlapped samples among these phenotypes.}

\item{genoMethod}{Name of methods to generate Genotypes, which are "SNPfrequency",
or "Haplotype". Default is "SNPfrequency".}

\item{SNPfrequency}{Vector of allele frequencies [double] from which to sample.}

\item{cSNP}{Proportion [double] of causal SNPs; used as genetic variant effects.
default is 0.01.}

\item{genVar}{[K x K] matrix of genetic covariance between any pair of phenotypes.}

\item{noiseVar}{[K x K] matrix of noise covaraince between any pair od phenotypes.}

\item{pcorr}{[K x K] matrix of correlation [double] between phenotypes. If input
is a single value which means the correlation between any pair of
phenotypes are same. Default is 0.5.}

\item{pgenVar}{Proportion of phenotype correlation influenced by genetic
variants. Default is 0.6. It is used with \code{pcorr} parameters.}

\item{h2s}{K by 1 vector of genomic heritability of different phenotypes.
Each element means the proportion of variance of a phenotype that can be
explained by a linear regression on a set of variants. Default is 0.5.}

\item{theta}{Proportion [double] of variance of shared genetic variant effects.}

\item{pTraits}{proportion [double] of traits which are correlated to each other.
Allows to simulate for instance different levels of pleiotropy. Using
ceiling(K*pTraits) as the number of phenotypes correlated to each other.}

\item{Haplotype}{Haplotype pool (two objects, first is the hoplotypes,
second object is the SNP information)}
}
\value{
Named list of i) the final simulated phenotype components
  (phenoComponentsFinal), ii) a named list of raw components (rawComponents)
  used for genetic effect simulation (genotypes and/or kinship, eigenvalues
  and eigenvectors of kinship)
}
\description{
Phenotypes are modeled under a linear additive model where Y = GB + E,
GB the genetic variant effects, E is noise term.
}
\examples{
K <- 5; N<- c(100,200,300,150,200); M<-160; Ns<-100


}
