% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simulateDist}
\alias{simulateDist}
\alias{probGen2expGen}
\alias{expGen2probGen}
\title{Data simulation for different distributions.}
\usage{
simulateDist(
  x,
  dist = c("unif", "norm", "bin", "cat_norm", "cat_unif"),
  m = NULL,
  std = 1,
  categories = NULL,
  prob = NULL
)

probGen2expGen(probGeno)

expGen2probGen(geno)
}
\arguments{
\item{x}{The number [integer] of observations to simulate.}

\item{dist}{Name of distribution [string] from which the observations are
drawn. 'norm' is the normal distribution, 'unif' the uniform distribution
'bin' the binomial distribution, "cat_norm" samples categorical variables
according to a normal distribution and "cat_unif"
according to a uniform distribution. For "cat_norm", length(category)/2 is
used mean for the normal distribution unless
specified otherwise.}

\item{m}{Mean of the normal distribution [double]/the mean between min
and max for the uniform distribution [double]/
the rank of the category to be used as mean for "cat_norm" [integer].}

\item{std}{Standard deviation of the normal distribution or the distance
of min/max from the mean for the uniform distribution [double].}

\item{categories}{Number of categories [integer] for simulating categorical
variables (for distr="cat_norm" or "cat_unif").}

\item{prob}{Probability [double] of success for each trial
(for distr="bin").}

\item{probGeno}{Vector [numeric] with genotype probabilites; has to be a
multiple of 3.}

\item{geno}{Vector [numeric] with genotypes}
}
\value{
Numeric vector of length [x] with the sampled values

Numeric vector of length [length(probGeno)/3] with the expected
genotype value per individual.

Numeric vector of length [length(geno)*3] with the genotype encoded
as probabbilities (p(AA), p(Aa), p(aa)).
}
\description{
Wrapper function to simulate data from different distribution with different
parameter settings.

Convert genotypes encoded as triplets of probablities (p(AA), p(Aa), p(aa))
into their expected genotype frequencies by 0*p(AA) + p(Aa) + 2p(aa).

Convert genotype frequencies to genotypes encoded as triplets of probablities
(p(AA), p(Aa), p(aa)).
}
\examples{
normal <- simulateDist(x=10, dist="norm", m=2, std=4)
cat_normal <- simulateDist(x=10, dist="cat_norm", categories=5)
cat_uniform <- simulateDist(x=10, dist="cat_unif", categories=5)
uniform <- simulateDist(x=10, dist="unif", m=4, std=1)
binomial <- simulateDist(x=10, dist="bin", prob=0.4)
nrSamples <- 10
# Construct genotype probability vector (usually from external input)
# First, assign zero probabilty of AA, Aa and aa for all samples
genotype_prob <- rep(0, 3*nrSamples)
# Second, for each sample draw one of 0,1,2 (corresponding to AA, Aa and aa)
genotype_prob[seq(1, nrSamples*3, 3) + sample(0:2, 10, replace=TRUE)] <- 1
genotype_exp <- probGen2expGen(genotype_prob)
nrSamples <- 10
# Simulate binomial SNP with 0.2 allele frequency
geno <- rbinom(nrSamples, 2, p=0.2)
geno_prob<- expGen2probGen(geno)
}
\seealso{
\code{\link{runif}}, \code{\link{rnorm}}, \code{\link{rbinom}} for
documentation of the underlying distributions.
}
